requirements.txt
Pillow>=10.0.0
matplotlib>=3.8.0
PyMuPDF>=1.24.0
